Fix all app-side compiler warnings
- Fixed conversion warnings from size_t to int32_t by using static_cast
- Fixed conversion warnings from sgx_enclave_id_t (uint64_t) to uint32_t in debug_dump calls
- Fixed unused variable warnings by using conditional compilation for debug-only variables
- Fixed C++17 structured bindings warnings by using C++11 compatible code with pairs
- Fixed double to int32_t conversion warnings with explicit casts
- Fixed unused parameter warning by commenting out parameter name
- Reduced total warnings from 50 to 12 (remaining are all in enclave code)
- All tests pass: two_center_chain and TB1

Also reorganized memory files (CLAUDE.md and CLAUDE.local.md) to have clear RULES sections at the top for better visibility of critical rules.

+++ impl/src/app/algorithms/align_concat.cpp
@@ -137,12 +137,12 @@ Table AlignConcat::AlignAndConcatenate(const Table& accumulator,
     
     // Dump sorted accumulator (left side of concatenation)
     debug_dump_table(result, ("before_concat_accumulator_" + concat_label).c_str(), 
-                    ("align_step5a_before_concat_" + concat_label).c_str(), eid,
+                    ("align_step5a_before_concat_" + concat_label).c_str(), static_cast<uint32_t>(eid),
                     {META_INDEX, META_ORIG_IDX, META_LOCAL_MULT, META_FINAL_MULT, META_FOREIGN_SUM, META_COPY_INDEX, META_ALIGN_KEY, META_JOIN_ATTR}, true);
     
     // Dump aligned child (right side of concatenation)  
     debug_dump_table(aligned_child, ("before_concat_child_" + concat_label).c_str(),
-                    ("align_step5b_before_concat_" + concat_label).c_str(), eid,
+                    ("align_step5b_before_concat_" + concat_label).c_str(), static_cast<uint32_t>(eid),
                     {META_INDEX, META_ORIG_IDX, META_LOCAL_MULT, META_FINAL_MULT, META_FOREIGN_SUM, META_COPY_INDEX, META_ALIGN_KEY, META_JOIN_ATTR}, true);
     
     DEBUG_INFO("Table sizes - Accumulator: %zu rows, Child: %zu rows", result.size(), aligned_child.size());
@@ -167,7 +167,7 @@ Table AlignConcat::AlignAndConcatenate(const Table& accumulator,
     
     // Debug: Dump final result - show concatenated attributes WITH ALL COLUMNS
     debug_dump_table(result, ("final_result_" + concat_label).c_str(), 
-                    ("align_step5_" + concat_label).c_str(), eid,
+                    ("align_step5_" + concat_label).c_str(), static_cast<uint32_t>(eid),
                     {META_INDEX, META_ORIG_IDX, META_LOCAL_MULT, META_FINAL_MULT, META_FOREIGN_SUM, META_COPY_INDEX, META_ALIGN_KEY, META_JOIN_ATTR}, true);
     
     DEBUG_INFO("========================================");
